+++ spec/unit/alf-engine/leaf/test_each.rb
@@ -1,0 +1,24 @@
+require 'spec_helper'
+module Alf
+  module Engine
+    describe Leaf, 'each' do
+
+      let(:input) {[
+        {"foo" => "bar"},
+      ]}
+
+      let(:leaf){
+        Leaf.new(input)
+      }
+
+      subject{ leaf.each.to_a }
+
+      let(:expected) {[
+        {foo: "bar"}
+      ]}
+
+      it{ should eq(expected) }
+
+    end
+  end
+end